* Fix homunc & code cleanup [Toms]   - Timer problems on delete_timer   - Intimacy problem (overflow & new values)   - Homunc deleted if intimacy < 0   - base exp is now given to master   - Homunc sometimes not saved
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@7913 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/mmo.h
@@ -174,7 +174,7 @@ struct s_homunculus {	//[orn]
 	int char_id;
 	short class_;
 	int hp,max_hp,sp,max_sp;
-	int intimacy;	//[orn]
+	unsigned long intimacy;	//[orn]
 	short hunger;
 	struct skill hskill[MAX_HOMUNSKILL]; //albator
 	short skillpts;
